\name{expectedList}
\alias{expectedList}
\title{Expected list.}
\description{Converts a \code{coin} object to a links data frame with coincidences and expected values.}
\usage{
expectedList(data, names = NULL, min = 1)
}
\arguments{
  \item{data}{is a \code{coin} object. See \code{\link{coin}}}
  \item{names}{a character vector.}
  \item{min}{minimum value of the statistic to include the edge in the list.}
}
\value{A links data frame with coincidences and expected values.}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{http://casus.usal.es/blog/modesto-escobar/}
}
