\name{network.addLayout}
\alias{network.addLayout}
\title{Produce interactive network.}
\author{
David Barrios and Carlos Prieto. Bioinformatics Service of Nucleus, University of Salamanca. See \url{http://bioinfo.usal.es/}
}
\value{
Return a network object.
}

\usage{
network.addLayout(net,layout)
}
\arguments{
\item{net}{a network object.}

\item{layout}{a matrix with the x and y coordinates for each node.}
}
\description{
\code{network.addLayout} add images to an interactive network.
}
