\name{network.addLinkFilter}
\alias{network.addLinkFilter}
\title{Produce interactive network.}
\author{
David Barrios and Carlos Prieto. Bioinformatics Service of Nucleus, University of Salamanca. See \url{http://bioinfo.usal.es/}
}
\value{
Return a network object.
}

\usage{
network.addLinkFilter(net, ...)
}
\arguments{
\item{net}{a network object.}

\item{...}{conditions for filtering links.}
}
\description{
\code{network.addLinkFilter} add node filters to an interactive network.
}
