\name{network.addNodeFilter}
\alias{network.addNodeFilter}
\title{Produce interactive network.}
\author{
David Barrios and Carlos Prieto. Bioinformatics Service of Nucleus, University of Salamanca. See \url{http://bioinfo.usal.es/}
}
\value{
Return a network object.
}

\usage{
network.addNodeFilter(net, ...)
}
\arguments{
\item{net}{a network object.}

\item{...}{conditions for filtering nodes.}
}
\description{
\code{network.addNodeFilter} add node filters to an interactive network.
}
