\name{network.nodePie}
\alias{network.nodePie}
\title{Produce interactive network.}
\author{
David Barrios and Carlos Prieto. Bioinformatics Service of Nucleus, University of Salamanca. See \url{http://bioinfo.usal.es/}
}
\value{
Return a network object.
}

\usage{
network.nodePie(net, pie)
}
\arguments{
\item{net}{a network object.}

\item{pie}{a numeric matrix with the rows in the same order as in the nodes data frame.}
}
\description{
\code{network.nodePie} add pie charts to an interactive network.
}
